requirements.txt
python-pptx
googletrans==4.0.0rc1
deepl
python-dotenv